+++ examples/survey/012c_rbend_line.py
@@ -7,76 +7,11 @@ from survey_utils import (
     misalignment_from_absolute_position,
     plot_exy,
     plot_exz,
+    rst_endpoint_offsets_from_parameters,
+    rst_from_reference_start,
 )
 
 elements_to_process = ['bend_1', 'bend_2', 'bend_3', 'q1', 'q2']
-
-
-def rst_from_reference_start(
-        XYZ_ref_start, E_ref_start, rot_s_rad, angle):
-    p_ref_start = np.eye(4)
-    p_ref_start[:3, :3] = E_ref_start
-    p_ref_start[:3, 3] = XYZ_ref_start
-
-    p_rst_start = MADPoint(p_ref_start)
-    p_rst_start.rpsi(rot_s_rad)
-    p_rst_start.rtheta(-angle / 2)
-
-    # S is along the chord, T is normal to the curvature plane, and R = S x T.
-    es = p_rst_start.matrix[:3, 2]
-    et = p_rst_start.matrix[:3, 1]
-    er = np.cross(es, et)
-
-    E_rst_start = np.column_stack((er, es, et))
-    return p_rst_start.matrix[:3, 3].copy(), E_rst_start
-
-
-def rst_endpoint_offsets_from_parameters(element, length):
-    angle = getattr(element, 'angle', 0.0)
-
-    rotation_tilt = MADPoint.psi_matrix(
-        element.rot_s_rad)[:3, :3]
-    rotation_half_angle = MADPoint.theta_matrix(
-        -angle / 2)[:3, :3]
-    rotation_misalignment = (
-        MADPoint.theta_matrix(element.rot_y_rad)
-        @ MADPoint.phi_matrix(element.rot_x_rad)
-        @ MADPoint.psi_matrix(element.rot_s_rad_no_frame)
-    )[:3, :3]
-
-    # E_rst = E_ref @ rotation_tilt @ rotation_half_angle @ xys_from_rst.
-    # The columns of xys_from_rst are R=-x, S=s, T=y, expressed in the
-    # tilted chord frame.
-    xys_from_rst = np.array([
-        [-1.0, 0.0, 0.0],
-        [0.0, 0.0, 1.0],
-        [0.0, 1.0, 0.0],
-    ])
-    rst_from_xys = (
-        xys_from_rst.T
-        @ rotation_half_angle.T
-        @ rotation_tilt.T
-    )
-
-    # The screenshot uses (DX, DS, DY). Xtrack uses (x, y, s), hence the
-    # corresponding vectors below are (DX, DY, DS) and (0, 0, l_E).
-    displacement_E_xys = np.array([
-        element.shift_x,
-        element.shift_y,
-        element.shift_s,
-    ])
-    body_chord_xys = np.array([0.0, 0.0, length])
-
-    b_E = rst_from_xys @ displacement_E_xys
-    b_S = rst_from_xys @ (
-        displacement_E_xys
-        + rotation_misalignment
-        @ rotation_tilt
-        @ rotation_half_angle
-        @ body_chord_xys
-    )
-    return b_E, b_S
-
 
 env = xt.Environment()
 env.set_particle_ref('proton', p0c=400e9)
@@ -135,7 +70,6 @@ line_no_jumps = line.copy(shallow=False)
 line_no_jumps['translation'].shift_x = 0
 line_no_jumps['rotation'].rot_y_rad = 0
 
-
 for elem_name in elements_to_process:
     ee_nj = line_no_jumps[elem_name]
 
@@ -150,6 +84,31 @@ for elem_name in elements_to_process:
     ee_nj.rot_shift_anchor = 0
 
 sv0_nj = line_no_jumps.survey(include_element_frames=True)
+tt0_nj = line_no_jumps.get_table(attr=True)
+
+XYZ_rst_start = []
+E_rst_start = []
+for ii in range(len(sv0_nj)):
+    XYZ_rst, E_rst = rst_from_reference_start(
+        XYZ_ref_start=sv0_nj.XYZ_ref_start[ii],
+        E_ref_start=sv0_nj.E_ref_start[ii],
+        rot_s_rad=tt0_nj.rot_s_rad[ii],
+        angle=tt0_nj.angle[ii],
+    )
+    XYZ_rst_start.append(XYZ_rst)
+    E_rst_start.append(E_rst)
+
+    chord = sv0_nj.XYZ_ref_end[ii] - sv0_nj.XYZ_ref_start[ii]
+    chord_length = np.linalg.norm(chord)
+    if chord_length > 0:
+        chord /= chord_length
+        xo.assert_allclose(E_rst[:, 1], chord, atol=1e-12, rtol=0)
+
+tt_rst = xt.Table({
+    'name': sv0_nj.name,
+    'XYZ': np.array(XYZ_rst_start),
+    'E': np.array(E_rst_start),
+})
 
 for elem_name in elements_to_process:
     ee_nj = line_no_jumps[elem_name]
@@ -186,32 +145,6 @@ for elem_name in elements_to_process:
 
 sv_nj = line_no_jumps.survey(include_element_frames=True)
 tt_nj = line_no_jumps.get_table(attr=True)
-
-XYZ_rst_start = []
-E_rst_start = []
-for elem_name in elements_to_process:
-    ee_nj = line_no_jumps[elem_name]
-    XYZ_rst, E_rst = rst_from_reference_start(
-        XYZ_ref_start=sv_nj['XYZ_ref_start', elem_name],
-        E_ref_start=sv_nj['E_ref_start', elem_name],
-        rot_s_rad=ee_nj.rot_s_rad,
-        angle=getattr(ee_nj, 'angle', 0.0),
-    )
-    XYZ_rst_start.append(XYZ_rst)
-    E_rst_start.append(E_rst)
-
-    chord = (
-        sv_nj['XYZ_ref_end', elem_name]
-        - sv_nj['XYZ_ref_start', elem_name]
-    )
-    chord /= np.linalg.norm(chord)
-    xo.assert_allclose(E_rst[:, 1], chord, atol=1e-12, rtol=0)
-
-tt_rst = xt.Table({
-    'name': np.array(elements_to_process),
-    'XYZ': np.array(XYZ_rst_start),
-    'E': np.array(E_rst_start),
-})
 
 b_E_rst = []
 b_S_rst = []
@@ -256,27 +189,12 @@ tt_misalignment_rst = xt.Table({
     'l_E': np.array(element_lengths),
 })
 
-XYZ_rst_start_all = []
-E_rst_start_all = []
-for ii in range(len(sv_nj)):
-    XYZ_rst, E_rst = rst_from_reference_start(
-        XYZ_ref_start=sv_nj.XYZ_ref_start[ii],
-        E_ref_start=sv_nj.E_ref_start[ii],
-        rot_s_rad=tt_nj.rot_s_rad[ii],
-        angle=tt_nj.angle[ii],
-    )
-    XYZ_rst_start_all.append(XYZ_rst)
-    E_rst_start_all.append(E_rst)
-
-XYZ_rst_start_all = np.array(XYZ_rst_start_all)
-E_rst_start_all = np.array(E_rst_start_all)
-
-displacement_start = sv.XYZ_elem_start - XYZ_rst_start_all
-displacement_end = sv.XYZ_elem_end - XYZ_rst_start_all
+displacement_start = sv.XYZ_elem_start - tt_rst.XYZ
+displacement_end = sv.XYZ_elem_end - tt_rst.XYZ
 offset_start_rst = np.einsum(
-    'nij,ni->nj', E_rst_start_all, displacement_start)
+    'nij,ni->nj', tt_rst.E, displacement_start)
 offset_end_rst = np.einsum(
-    'nij,ni->nj', E_rst_start_all, displacement_end)
+    'nij,ni->nj', tt_rst.E, displacement_end)
 
 supports_misalignment = np.zeros(len(sv_nj), dtype=bool)
 for ii, element in enumerate(line_no_jumps.elements):
@@ -316,7 +234,7 @@ tt_align = xt.Table({
     'bgamma': -tt_nj['rot_s_rad_no_frame'],
 
     # RST unit vectors at element start
-    'E_rst_start': E_rst_start_all,
+    'E_rst_start': tt_rst.E,
 })
 
 tt_align.to_tfs('test_align.tfs')
